rename ExampleConfig.hpp => Config.hpp
